Extract scheduled payouts into a dedicated admin controller
ScheduledPayout actions don't belong on PayoutsController (different model, awkward scheduled_* prefixes). The new controller exposes RESTful routes — GET /internal/admin/scheduled_payouts and POST /scheduled_payouts/:id/{execute,cancel} — and the actions drop the redundant prefix. Audit log action strings renamed accordingly (payouts.scheduled_execute -> scheduled_payouts.execute, ditto cancel). The bundled scheduled_payouts in the user payouts index is unchanged.
