+++ config/routes.rb
@@ -346,11 +346,13 @@ Rails.application.routes.draw do
               post :pause
               post :resume
               post :issue
-              post :scheduled_list
             end
+          end
+
+          resources :scheduled_payouts, only: [:index] do
             member do
-              post :scheduled_execute
-              post :scheduled_cancel
+              post :execute
+              post :cancel
             end
           end
